chore: rename price_feed to price_tick
